chore: remove meteor.startup from follow message
