+++ apps/meteor/app/threads/client/index.ts
@@ -1,3 +1,2 @@
-import './messageAction/follow';
 import './messageAction/replyInThread';
 import './threads.css';
